test(vault): split pr_a18_test.go by concern + sweep PR-id from comments
Test-organisation cleanup following PR-A18 (#240):

- Delete adapters/vault/pr_a18_test.go (348 LOC, named after PR id which
  violates Go test-file convention) and split into three concern-named
  files reusing the shared fakeVaultClient:
    - transit_namespace_test.go (44 LOC, applyNamespaceFromEnv)
    - transit_datakey_test.go (66 LOC, datakey envelope encrypt path)
    - transit_version_cache_test.go (141 LOC, lock-free atomic.Int64 cache)
- Drop the duplicate datakeyFake (the shared fakeVaultClient already
  routes /datakey/plaintext after PR#240) and the handwritten itoa()
  helper. The shared fake gains a separate datakeyCalls counter so the
  "encryptCalls must stay 0" regression guard for the lock-free encrypt
  flow stays sharp.
- Add a readCalls counter to fakeVaultClient.Read so the Rotate
  invalidation tests can assert "no Vault read on cache hit" precisely.
- Sweep PR-A18 references out of code comments (kept only where the
  comment anchors a meaningful boundary, e.g. doc.go decrypt forward-
  compat). Project rule: PR ids belong in the PR description, not in
  long-lived code comments.

No production code changes. Build / -race / -tags=integration / lint
all green.
